sqlite compatible experimental Signed-off-by: Juan Iscar <juan.iscar@exponentiateam.com>
