+++ go.mod
@@ -5,6 +5,7 @@ go 1.14
 require (
 	github.com/jinzhu/gorm v1.9.12
 	github.com/lib/pq v1.1.1
+	github.com/mattn/go-sqlite3 v2.0.1+incompatible
 	github.com/onsi/ginkgo v1.8.0
 	github.com/onsi/gomega v1.5.0
 )
